Fixes #9: case IDs added to reports
