+++ src/waste/cli.py
@@ -2,8 +2,6 @@ from pathlib import Path
 
 import click
 
-from . import transportation as tp
-
 
 @click.group()
 def main():
@@ -19,7 +17,9 @@ def main():
 @click.option('-p', '--parallel', is_flag=True, default=True, show_default=True,
               help='Run the tool using all available cores in parallel.')
 def transportation(log_path: Path, output_dir: Path, parallel: bool):
-    result = tp.identify(log_path, parallel)
+    from waste import transportation
+
+    result = transportation.identify(log_path, parallel)
 
     output_dir.mkdir(parents=True, exist_ok=True)
