+++ tests/assets/bimp-example_case_handoff_5.csv
@@ -1,6 +1,6 @@
-,source_activity,source_resource,destination_activity,destination_resource,duration,frequency
-0,Assess application,Credit Officer-000002,Notify rejection,Credit Officer-000001,0 days 00:19:13.845001,1.0
-1,Check income sources,Clerk-000001,Assess application,Credit Officer-000002,0 days 01:07:31.947001,1.0
-2,Credit application received,NA,Check income sources,Clerk-000001,0 days 00:16:24.456001,1.0
-3,Notify rejection,Credit Officer-000001,Receive customer feedback,System-000001,0 days 00:00:00.000001,1.0
-4,Receive customer feedback,System-000001,Credit  application  processed,NA,0 days 00:00:00.000001,1.0
+,source_activity,source_resource,destination_activity,destination_resource,duration,frequency,case_id
+0,Assess application,Credit Officer-000002,Notify rejection,Credit Officer-000001,0 days 00:19:13.845001,1.0,1
+1,Check income sources,Clerk-000001,Assess application,Credit Officer-000002,0 days 01:07:31.947001,1.0,1
+2,Credit application received,NA,Check income sources,Clerk-000001,0 days 00:16:24.456001,1.0,1
+3,Notify rejection,Credit Officer-000001,Receive customer feedback,System-000001,0 days 00:00:00.000001,1.0,1
+4,Receive customer feedback,System-000001,Credit  application  processed,NA,0 days 00:00:00.000001,1.0,1
