+++ tests/assets/bimp-example_case_handoff_6.csv
@@ -1,7 +1,7 @@
-,source_activity,source_resource,destination_activity,destination_resource,duration,frequency
-0,Assess application,Credit Officer-000001,Make credit offer,Credit Officer-000002,0 days 00:20:54.142001,1.0
-1,Make credit offer,Credit Officer-000002,Credit  application  processed,NA,0 days 00:00:00.000001,1.0
-2,Credit application received,NA,Check credit history,Clerk-000002,0 days 00:00:00.000001,1.0
-3,Credit application received,NA,Check income sources,Clerk-000001,0 days 00:01:56.542001,1.0
-4,Check credit history,Clerk-000002,Assess application,Credit Officer-000001,0 days 01:10:27.969001,1.0
-5,Check income sources,Clerk-000001,Assess application,Credit Officer-000001,0 days 00:55:27.209001,1.0
+,source_activity,source_resource,destination_activity,destination_resource,duration,frequency,case_id
+0,Assess application,Credit Officer-000001,Make credit offer,Credit Officer-000002,0 days 00:20:54.142001,1.0,1
+1,Make credit offer,Credit Officer-000002,Credit  application  processed,NA,0 days 00:00:00.000001,1.0,1
+2,Credit application received,NA,Check credit history,Clerk-000002,0 days 00:00:00.000001,1.0,1
+3,Credit application received,NA,Check income sources,Clerk-000001,0 days 00:01:56.542001,1.0,1
+4,Check credit history,Clerk-000002,Assess application,Credit Officer-000001,0 days 01:10:27.969001,1.0,1
+5,Check income sources,Clerk-000001,Assess application,Credit Officer-000001,0 days 00:55:27.209001,1.0,1
